Added music
Added music to the game, it repeats when the song is over. I just
hardcoded the mediaplayer stop time to the duration of the song.
PaintPanel checks if mediaplayer's currentTime is == songs hardcoded
duration, and if it is, mediaPlayer seeks to time 0, so song starts
over. Volume slider in options also works.